Fix wrong half-float usage in Cycles X
Was generating strict compiler warning.

Probably not measurable for users since this was not code path
which is executed yet.

+++ intern/cycles/integrator/pass_accessor.cpp
@@ -127,7 +127,7 @@ static void pad_pixels(const BufferParams &buffer_params,
   }
 
   if (destination.pixels_half_rgba) {
-    const float one = float_to_half(1.0f);
+    const half one = float_to_half(1.0f);
     half4 *pixel = destination.pixels_half_rgba;
 
     for (size_t i = 0; i < size; i++, pixel++) {
